Moved checkHeapSizeLimit to runtime.ts
